restructured crud functionalilty of projects
