+++ assets/recurring-elements/nav.css
@@ -71,9 +71,9 @@ nav {
     width: 100%;
     background-color: #696969;
     /* Dark gray */
-    padding: 1em 0;
-    box-shadow: 0 -2px 4px rgba(0, 0, 0, 0.1);
-    z-index: 1000;
+    padding: 0.8em 0;
+    box-shadow: -2px 4px rgba(0, 0, 0, 0.4);
+    z-index: 0.1;
     text-align: center;
 }
