+++ assets/src/projects/index.html
@@ -1,16 +1,19 @@
 <!DOCTYPE html>
 <html>
+
 <head>
     <title>Projects Page</title>
-    <link href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap"
+        rel="stylesheet">
     <link rel="stylesheet" href="style.css">
     <link rel="stylesheet" href="../../recurring-elements/nav.css">
-    <script defer src="importProjects.js"></script>
-    <script defer src="addTasks.js"></script>
+    <script type="module" src="importProjects.js"></script>
     <script defer src="../../recurring-elements/nav.js"></script>
     <script defer src="../../recurring-elements/footer.js"></script>
-    <link href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css2?family=Josefin+Sans:ital,wght@0,100..700;1,100..700&display=swap"
+        rel="stylesheet">
 </head>
+
 <body>
     <header>
         <div id="insert-nav" title="Projects"></div>
@@ -23,4 +26,5 @@
     </main>
     <div id="footer"></div>
 </body>
+
 </html>
